An initial attempt at a modern project format
I have a big problem with how Composer installs dependencies,
dragging in everything in the repositroy without a means to define
what files are only needed for distribution.

There's too many levels of abstraction and bloat and I'm not a fan of these
requirements, but it is probably necessary for the code to be seen and used
by others.